+++ src/CfdiUtils/Elements/Nomina12/HorasExtra.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace CfdiUtils\Elements\Nomina12;
+
+use CfdiUtils\Elements\Common\AbstractElement;
+
+class HorasExtra extends AbstractElement
+{
+    public function getElementName(): string
+    {
+        return 'nomina12:HorasExtra';
+    }
+}