+++ tests/CfdiUtilsTests/Elements/Nomina12/HorasExtraTest.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace CfdiUtilsTests\Elements\Nomina12;
+
+use CfdiUtils\Elements\Nomina12\HorasExtra;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @covers \CfdiUtils\Elements\Nomina12\HorasExtra
+ */
+class HorasExtraTest extends TestCase
+{
+    /** @var HorasExtra */
+    public $element;
+
+    protected function setUp()
+    {
+        parent::setUp();
+        $this->element = new HorasExtra();
+    }
+
+    public function testConstructedObject()
+    {
+        $this->assertSame('nomina12:HorasExtra', $this->element->getElementName());
+    }
+}